Final tweaks for release: Add warning for the new config option, change maxHealth to ints

+++ src/main/java/ichttt/mods/firstaid/damagesystem/DamageablePart.java
@@ -16,8 +16,8 @@ import java.util.Arrays;
 
 public class DamageablePart implements INBTSerializable<NBTTagCompound> {
 
-    public final float initialMaxHealth;
-    private float maxHealth;
+    public final int initialMaxHealth;
+    private int maxHealth;
     public final boolean canCauseDeath;
     @Nullable
     public PartHealer activeHealer;
@@ -29,7 +29,7 @@ public class DamageablePart implements INBTSerializable<NBTTagCompound> {
     public float currentHealth;
     private float absorption;
 
-    public DamageablePart(float maxHealth, boolean canCauseDeath, @Nonnull EnumPlayerPart playerPart, @Nonnull IDebuff... debuffs) {
+    public DamageablePart(int maxHealth, boolean canCauseDeath, @Nonnull EnumPlayerPart playerPart, @Nonnull IDebuff... debuffs) {
         this.initialMaxHealth = maxHealth;
         this.maxHealth = maxHealth;
         this.canCauseDeath = canCauseDeath;
@@ -120,12 +120,12 @@ public class DamageablePart implements INBTSerializable<NBTTagCompound> {
         return absorption;
     }
 
-    public void setMaxHealth(float maxHealth) {
+    public void setMaxHealth(int maxHealth) {
         this.maxHealth = Math.max(2, maxHealth); //set 2 as a minimum
         this.currentHealth = Math.min(currentHealth, this.maxHealth);
     }
 
-    public float getMaxHealth() {
+    public int getMaxHealth() {
         return maxHealth;
     }
 }
